icecream함수 icecream.swift로 분리
