+++ CodeStarterCamp_Week1/pepero.swift
@@ -1,0 +1,8 @@
+//
+//  pepero.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by Won bo Shim on 2022/04/16.
+//
+
+import Foundation
